M2-6493: revert ParticipantActivityDetails
- Handle updated DataViz in RespondentData and RespondentDataFeature
- Add updated Header to DataViz
