--- src/modules/Dashboard/features/ParticipantActivity/Details/HeaderOptions/HeaderOptions.types.ts
@@ -1,5 +1,0 @@
-export interface HeaderOptionsProps {
-  dataTestid: string;
-  onTakeNow: () => void;
-  onAssignActivity: () => void;
-}
