--- src/modules/Dashboard/features/ParticipantActivity/Details/HeaderOptions/index.ts
@@ -1,1 +1,0 @@
-export * from './HeaderOptions';
