--- src/modules/Dashboard/features/ParticipantActivity/Details/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ParticipantActivityDetails';
