--- src/modules/Dashboard/features/ParticipantActivity/Responses/Responses.tsx
@@ -1,5 +1,0 @@
-import RespondentDataReview from '../../RespondentData/RespondentDataReview';
-
-export const ParticipantActivityResponses = () => <RespondentDataReview />;
-
-export default ParticipantActivityResponses;
