--- src/modules/Dashboard/features/ParticipantActivity/Responses/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Responses';
