--- src/modules/Dashboard/features/ParticipantActivity/Summary/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Summary';
