--- src/modules/Dashboard/features/ParticipantActivity/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ParticipantActivity';
