+++ src/modules/Dashboard/features/RespondentData/RespondentDataHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './RespondentDataHeader';
