+++ src/modules/Dashboard/pages/RespondentData/RespondentData.utils.ts
@@ -1,0 +1,12 @@
+import { Roles } from 'shared/consts';
+
+export const hasPermissionToViewData = (roles?: Roles[] | null) => {
+  if (!roles?.length) return false;
+
+  return (
+    roles.includes(Roles.SuperAdmin) ||
+    roles.includes(Roles.Owner) ||
+    roles.includes(Roles.Manager) ||
+    roles.includes(Roles.Reviewer)
+  );
+};
